keep up with thegood  work!

+++ src/controllers/empleado.controller.js
@@ -1,7 +1,6 @@
 import { devLogger } from '../config/logs/logger.config.js';
 import { EmpleadoDTO } from '../services/db/dto/empleado.dto.js';
-import { equipoInformaticoService } from '../services/repository/services.js';
-import { empleadoService } from '../services/repository/services.js';
+import { empleadoService, equipoInformaticoService, oficinaService } from '../services/repository/services.js';
 
 export async function crearEmpleado(req, res) {
   const obj = req.body; try {
@@ -78,8 +77,13 @@ export async function actualizarEmpleado(req, res) {
 export async function eliminarEmpleado(req, res) {
   const empleadoId = req.params.id;
   try {
-    await empleadoService.deleteEmpleado(empleadoId);
-    return res.sendSuccess({ state: "Employee deleted" });
+    const empleado = await empleadoService.getEmpleadoById(empleadoId, { include: 'EquipoInformaticos' });
+    if (empleado.EquipoInformaticos.length > 0) {
+      res.sendClientError('Employee has hardware assigned');
+    } else {
+      await empleadoService.deleteEmpleado(empleadoId);
+      return res.sendSuccess({ state: "Employee deleted" });
+    }
   } catch (error) {
     devLogger.error(error);
     return res.sendInternalServerError(error);
@@ -103,6 +107,41 @@ export async function asignarEquipoAempleado(req, res) {
     return res.sendInternalServerError(error);
   }
 };
+
+export async function asignarOficinaAempleado(req, res) {
+    const empleadoId = req.params.empleadoId;
+    const oficinaId = req.params.oficinaId;
+    try {
+      const empleado = await empleadoService.getEmpleadoById(empleadoId, { include: 'Oficina' });
+      const oficina = await oficinaService.getOficinaById(oficinaId);
+      if (empleado.Oficina) {
+        res.sendClientError('office is already assigned to another employee');
+      }
+      empleado.setOficina(oficina);
+      res.sendSuccess({ state: "office assigned to employee" });
+    } catch (error) {
+      devLogger.error(error);
+      return res.sendInternalServerError(error);
+    }
+}
+
+export async function removerEquipoAempleado(req, res) {
+  const empleadoId = req.params.empleadoId;
+  const equipoInformaticoId = req.params.equipoId;
+  try {
+    const empleado = await empleadoService.getEmpleadoById(empleadoId, { include: 'EquipoInformaticos' });
+    const equipoInformatico = await equipoInformaticoService.getEquipoInformaticoById(equipoInformaticoId);
+    if (!equipoInformatico.Empleado) {
+      res.sendClientError('hardware is not assigned to any employee');
+    }
+    await equipoInformaticoService.updateEquipoInformatico(equipoInformaticoId, { "estado": "DISPONIBLE" });
+    empleado.removeEquipoInformaticos(equipoInformatico);
+    res.sendSuccess({ state: "hardware removed from employee" });
+  } catch (error) {
+    devLogger.error(error);
+    return res.sendInternalServerError(error);
+  }
+}
 
 export async function restaurarEmpleado(req, res) {
   const empleadoId = req.params.id;
